Refactor: Split EvergreenService and Lifecycle
