+++ src/main/java/com/aws/iot/evergreen/kernel/GenericExternalService.java
@@ -21,6 +21,7 @@ import java.util.function.IntConsumer;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import static com.aws.iot.evergreen.kernel.Lifecycle.TIMEOUT_NAMESPACE_TOPIC;
 import static com.aws.iot.evergreen.packagemanager.KernelConfigResolver.VERSION_CONFIG_KEY;
 
 @SuppressWarnings("PMD.NullAssignment")
@@ -86,7 +87,7 @@ public class GenericExternalService extends EvergreenService {
 
     @Override
     public void startup() throws InterruptedException {
-        RunStatus result = run(LIFECYCLE_STARTUP_NAMESPACE_TOPIC, exit -> {
+        RunStatus result = run(Lifecycle.LIFECYCLE_STARTUP_NAMESPACE_TOPIC, exit -> {
             runScript = null;
             if (getState() == State.INSTALLED) {
                 if (exit == 0) {
@@ -158,7 +159,6 @@ public class GenericExternalService extends EvergreenService {
         }
     }
 
-
     @Override
     @SuppressWarnings("PMD.CloseResource")
     public void shutdown() {
@@ -196,7 +196,7 @@ public class GenericExternalService extends EvergreenService {
      * @return the status of the run.
      */
     protected RunStatus run(String name, IntConsumer background) throws InterruptedException {
-        Node n = (getLifeCycleTopic() == null) ? null : getLifeCycleTopic().getChild(name);
+        Node n = (getLifecycleTopic() == null) ? null : getLifecycleTopic().getChild(name);
         return n == null ? RunStatus.NothingDone : run(n, background);
     }
